+++ __fixtures__/helper.js
@@ -1,14 +1,10 @@
-const log = (s)=> {
-  //if (s === )
-}
-
 module.exports = {
   mockContext: (options) => {
     if (!options) options = {}
 
     return {
       repo: (properties) => { return Object.assign({ owner: 'owner', repo: 'repo' }, properties) },
-      event: (options.event)? options.event : 'pull_request',
+      event: (options.event) ? options.event : 'pull_request',
       payload: {
         pull_request: {
           user: {
@@ -31,9 +27,9 @@ module.exports = {
         }
       },
       log: {
-        debug: (s)=> console.log(`TEST[debug] > ${JSON.stringify(s)}`),
-        info: (s)=> console.log(`TEST[info] > ${JSON.stringify(s)}`),
-        warn: (s)=> console.log(`TEST[warn] > ${JSON.stringify(s)}`)
+        debug: (s) => console.log(`TEST[debug] > ${JSON.stringify(s)}`),
+        info: (s) => console.log(`TEST[info] > ${JSON.stringify(s)}`),
+        warn: (s) => console.log(`TEST[warn] > ${JSON.stringify(s)}`)
       },
       github: {
         repos: {
@@ -101,7 +97,6 @@ module.exports = {
       }
     }
   },
-
 
   expectedStatus: (status, description) => {
     return {
