+++ lib/configuration.js
@@ -16,11 +16,11 @@ class Configuration {
     }
   }
 
-  isFlexVersion() {
-    return (process.env.MERGEABLE_VERSION === 'flex' && this.settings.version === 'flex')
+  isFlexVersion () {
+    return (process.env.MERGEABLE_VERSION === 'flex' && this.settings.version === 2)
   }
 
-  validate() {
+  validate () {
     if (this.settings.mergeable === undefined) {
       throw new Error(Configuration.ERROR_INVALID_YML)
     }
