+++ lib/handler.js
@@ -2,7 +2,6 @@ const Configuration = require('../lib/configuration')
 const fetch = require('node-fetch')
 const checks = require('../lib/checks')
 const stale = require('../lib/stale')
-const { isEventInContext } = require('../lib/validators/validator')
 
 class Handler {
   static async handleIssues (context) {
@@ -138,49 +137,6 @@ class Handler {
         {
           title: `An Error has occured, Please resolve them for mergeable to run properly`,
           summary: `Following Error has been reported from Mergeable:\n` + err.toString()})
-    })
-  }
-
-  // TODO Refactor this since it's responsibility goes beyond handling validators and actions.
-  // For now, a quick and easy way to have testable units through simple dependency injection.
-  static async handleFlex(context, registry) {
-    // TODO if PR do checks and verify.
-    let config = await Configuration.instanceWithContext(context)
-
-    // Quick and dirty for now. Registry should be as a singleton that caches instances
-    // of validators and actions - @jusx
-    if (registry === undefined) {
-      registry = { validators: new Map(), actions: new Map() }
-    }
-
-    context.log.debug('Processing: \n' + config.settings)
-    config.settings.mergeable.forEach((rule)=> {
-      let events = Array.isArray(rule.when)? rule.when : rule.when.split(',').map(n=>n.trim())
-      let executable
-
-      let promisses = []
-      rule.validate.forEach(validation => {
-        let key = validation.do
-        if (!registry.validators.has(key)) {
-          let Validator = require(`./validators/${key}`)
-          registry.validators.set(key, new Validator())
-        }
-
-        let validator = registry.validators.get(key)
-        let eventName = `${context.event}.${context.payload.action}`
-        if (isEventInContext(eventName, events) && validator.isEventSupported(eventName)) {
-          promisses.push(validator.validate(context, validation, config.settings))
-        }
-      })
-
-      Promise.all(promisses)
-        .then( results => {
-          context.log.debug(results)
-          // TODO call actions.
-        })
-        .catch(e => {
-          context.log.error(e)
-        })
     })
   }
 }
