+++ lib/mergeable.js
@@ -1,8 +1,7 @@
 const createScheduler = require('probot-scheduler')
 const Handler = require('./handler')
 const Configuration = require('../lib/configuration')
-const Title = require('../lib/title')
-
+const { isEventInContext } = require('../lib/validators/validator')
 
 class Mergeable {
   constructor (mode, version) {
@@ -32,7 +31,7 @@ class Mergeable {
   flex (robot) {
     robot.on('*', async context => {
       context.log.info({event: context.event, action: context.payload.action})
-      Handler.handleFlex(context)
+      processValidations(context)
     })
   }
 
@@ -64,4 +63,47 @@ class Mergeable {
   }
 }
 
-module.exports = Mergeable
+// TODO Refactor this since it's responsibility goes beyond handling validators and actions.
+// For now, a quick and easy way to have testable units through simple dependency injection.
+const processValidations = async (context, registry) => {
+  // TODO if PR do checks and verify.
+  let config = await Configuration.instanceWithContext(context)
+
+  // Quick and dirty for now. Registry should be as a singleton that caches instances
+  // of validators and actions - @jusx
+  if (registry === undefined) {
+    registry = { validators: new Map(), actions: new Map() }
+  }
+
+  context.log.debug('Processing: \n' + config.settings)
+  config.settings.mergeable.forEach((rule) => {
+    let events = Array.isArray(rule.when) ? rule.when : rule.when.split(',').map(n => n.trim())
+
+    let promisses = []
+    rule.validate.forEach(validation => {
+      let key = validation.do
+      if (!registry.validators.has(key)) {
+        let Validator = require(`./validators/${key}`)
+        registry.validators.set(key, new Validator())
+      }
+
+      let validator = registry.validators.get(key)
+      let eventName = `${context.event}.${context.payload.action}`
+      if (isEventInContext(eventName, events) && validator.isEventSupported(eventName)) {
+        promisses.push(validator.validate(context, validation, config.settings))
+      }
+    })
+
+    Promise.all(promisses)
+      .then(results => {
+        context.log.debug(results)
+        // TODO call actions for pass or fail
+      })
+      .catch(e => {
+        context.log.error(e)
+        // TODO call actions for error.
+      })
+  })
+}
+
+module.exports = { Mergeable: Mergeable, processValidations: processValidations }
