+++ lib/validators/label.js
@@ -1,7 +1,7 @@
 const { Validator } = require('./validator')
 
 class Label extends Validator {
-  constructor() {
+  constructor () {
     super()
     this.supportedEvents = [ 'pull_request.opened',
       'pull_request.edited',
@@ -15,7 +15,7 @@ class Label extends Validator {
     ]
   }
 
-  async validate(context, validationSettings) {
+  async validate (context, validationSettings) {
     let labels = await context.github.issues.getIssueLabels(
       context.repo({ number: this.getPayload(context) })
     )
