+++ lib/validators/title.js
@@ -1,7 +1,7 @@
 const { Validator } = require('./validator')
 
 class Title extends Validator {
-  constructor() {
+  constructor () {
     super()
     this.supportedEvents = [ 'pull_request.opened',
       'pull_request.edited',
@@ -15,7 +15,7 @@ class Title extends Validator {
     ]
   }
 
-  validate(context, validationSettings) {
+  validate (context, validationSettings) {
     return this.processOptions(
       validationSettings,
       this.getPayload(context).title
