+++ lib/validators/validator.js
@@ -1,11 +1,11 @@
 const processor = require('../processor')
 
 class Validator {
-  constructor() {
+  constructor () {
     this.processor = processor
   }
 
-  async validate() {
+  async validate () {
     return false
   }
 
@@ -16,26 +16,25 @@ class Validator {
    *
    * @return boolean true if the validator supports the event. i.e. issues.opened
    */
-  isEventSupported(eventName) {
+  isEventSupported (eventName) {
     return isEventInContext(eventName, this.supportedEvents)
   }
 
-  processOptions(vSettings, value) {
+  processOptions (vSettings, value) {
     // TODO pass in an array of supported validator option names
     return processor.processFilters(vSettings.do, value, vSettings)
   }
 
-  getPayload(context) {
+  getPayload (context) {
     return context.payload[context.event]
   }
 }
 
 const isEventInContext = (eventName, events) => {
   let eventObject = eventName.split('.')[0]
-  return ((events.indexOf(eventName.trim()) > -1)
-    || (events.indexOf('pull_request.*') > -1 && eventObject === 'pull_request')
-    || (events.indexOf('issues.*') > -1 && eventObject === 'issues'))
-
+  return ((events.indexOf(eventName.trim()) > -1) ||
+    (events.indexOf('pull_request.*') > -1 && eventObject === 'pull_request') ||
+    (events.indexOf('issues.*') > -1 && eventObject === 'issues'))
 }
 
 module.exports = {
